Move the code to a new module
