+++ magda-minion-framework/tsconfig.json
@@ -4,7 +4,10 @@
         "outDir": "./dist",
         "strictNullChecks": false,
         "rootDir": "./src",
-        "tsBuildInfoFile": "./tsconfig.tsbuildinfo"
+        "tsBuildInfoFile": "./tsconfig.tsbuildinfo",
+        "composite": true,
+        "declaration": true,
+        "declarationMap": true
     },
     "references": [{ "path": "../magda-typescript-common" }],
     "include": ["./src"]
