+++ magda-semantic-indexer-framework/src/chunker.ts
