+++ magda-semantic-indexer-framework/src/config.json
