+++ magda-semantic-indexer-framework/src/config.ts
