+++ magda-semantic-indexer-framework/src/index.ts
@@ -1,0 +1,24 @@
+export { default as SemanticIndexerOptions } from "./semanticIndexerOptions.js";
+export {
+    commonSemanticIndexerYargs,
+    SemanticIndexerArguments
+} from "./commonSemanticIndexerYargs.js";
+export {
+    ItemType,
+    SemanticIndexDocument,
+    buildSemanticIndexDocument
+} from "./indexSchema.js";
+export {
+    CreateEmbeddingText,
+    CreateEmbeddingTextParams,
+    EmbeddingText
+} from "./createEmbeddingText.js";
+export { Chunker, FixedLengthChunkStrategy } from "./chunker.js";
+export { SkipError } from "./skipError.js";
+export { indexEmbeddingText } from "./indexEmbeddingText.js";
+export { onRecordFoundRegistryRecord } from "./onRecordFoundRegistryRecord.js";
+export { onRecordFoundStorageObject } from "./onRecordFoundStorageObject.js";
+export { config } from "./config.js";
+
+import semanticIndexer from "./semanticIndexer.js";
+export default semanticIndexer;
