+++ magda-semantic-indexer-framework/src/indexSchema.ts
