+++ magda-semantic-indexer-framework/src/onRecordFoundStorageObject.ts
@@ -1,18 +1,18 @@
-import { onRecordFoundType } from "@magda/minion-sdk";
+import { onRecordFoundType } from "magda-minion-framework/src/MinionOptions.js";
 import { Chunker } from "./chunker.js";
-import EmbeddingApiClient from "../EmbeddingApiClient.js";
-import OpensearchApiClient from "../OpensearchApiClient.js";
+import EmbeddingApiClient from "magda-typescript-common/src/EmbeddingApiClient.js";
+import OpensearchApiClient from "magda-typescript-common/src/OpensearchApiClient.js";
 import SemanticIndexerOptions from "./semanticIndexerOptions.js";
 import { indexEmbeddingText } from "./indexEmbeddingText.js";
 import { EmbeddingText } from "./createEmbeddingText.js";
-import { Record } from "../generated/registry/api.js";
+import { Record } from "magda-typescript-common/src/generated/registry/api.js";
+import retry from "magda-typescript-common/src/retry.js";
 import fetch from "node-fetch";
 import { promises as fs } from "fs";
 import { tmpdir } from "os";
 import { join } from "path";
 import { v4 as uuidv4 } from "uuid";
 import { SkipError } from "./skipError.js";
-import retry from "../retry.js";
 
 export const onRecordFoundStorageObject = (
     userConfig: SemanticIndexerOptions,
