+++ magda-semantic-indexer-framework/src/semanticIndexerOptions.ts
