+++ magda-semantic-indexer-framework/src/skipError.ts
