+++ magda-semantic-indexer-framework/src/test/mockEmbeddingApi.ts
