+++ magda-semantic-indexer-framework/tsconfig.json
@@ -1,0 +1,11 @@
+{
+    "extends": "../tsconfig-global.json",
+    "compilerOptions": {
+        "outDir": "./dist",
+        "strictNullChecks": false,
+        "rootDir": "./src",
+        "tsBuildInfoFile": "./tsconfig.tsbuildinfo"
+    },
+    "include": ["./src"],
+    "references": [{ "path": "../magda-minion-framework" }]
+}
