+++ packages/semantic-indexer-sdk/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@magda/semantic-indexer-sdk",
   "description": "MAGDA Semantic Indexer SDK",
-  "version": "1.0.0",
+  "version": "5.2.0",
   "type": "module",
   "exports": {
     ".": "./dist/index.js",
@@ -23,16 +23,24 @@
   "author": "",
   "license": "Apache-2.0",
   "devDependencies": {
-    "@magda/typescript-common": "^5.2.0",
+    "@magda/semantic-indexer-framework": "^5.2.0",
     "@microsoft/api-extractor": "~7.39.0",
     "esbuild": "^0.19.10",
     "ts-loader": "^9.5.1",
     "typescript": "~5.3.3"
   },
+  "dependencies": {
+    "@magda/typescript-common": "^5.2.0",
+    "@opensearch-project/opensearch": "^3.5.1",
+    "@types/express": "^4.17.21",
+    "@types/urijs": "^1.19.19",
+    "@types/yargs": "^12.0.8"
+  },
   "magda": {
     "language": "typescript",
     "categories": {
-      "npmPackage": true
+      "npmPackage": true,
+      "useMinionLib": true
     }
   },
   "keywords": [
