+++ packages/semantic-indexer-sdk/src/index.ts
@@ -1,22 +1,3 @@
-export {
-    CreateEmbeddingText,
-    CreateEmbeddingTextParams,
-    EmbeddingText
-} from "@magda/typescript-common/dist/semantic-indexer/createEmbeddingText.js";
-
-export { default as semanticIndexer } from "@magda/typescript-common/dist/semantic-indexer/semanticIndexer.js";
-
-export { default as SemanticIndexerOptions } from "@magda/typescript-common/dist/semantic-indexer/semanticIndexerOptions.js";
-
-export {
-    commonSemanticIndexerYargs,
-    SemanticIndexerArguments
-} from "@magda/typescript-common/dist/semantic-indexer/commonSemanticIndexerYargs.js";
-
-export {
-    ItemType,
-    SemanticIndexDocument,
-    buildSemanticIndexDocument
-} from "@magda/typescript-common/dist/semantic-indexer/indexSchema.js";
-
-export { Record } from "@magda/typescript-common/dist/generated/registry/api.js";
+import semanticIndexer from "@magda/semantic-indexer-framework/dist/index.js";
+export * from "@magda/semantic-indexer-framework/dist/index.js";
+export default semanticIndexer;
